server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "anthropic>=0.75.0",
    "fastapi>=0.125.0",
    "langchain>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "pdfplumber>=0.11.8",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]
